Extract to dev-utils
