+++ packages/protocol/package.json
@@ -91,6 +91,7 @@
     "web3-provider-engine": "^15.0.0"
   },
   "devDependencies": {
+    "@celo/dev-utils": "0.0.1-dev",
     "@celo/flake-tracker": "0.0.1-dev",
     "@celo/typechain-target-web3-v1-celo": "0.2.0",
     "@celo/typescript": "0.0.1",
