+++ packages/protocol/scripts/release-report.ts
@@ -1,22 +1,4 @@
-import { concurrentMap } from '@celo/base'
-import fetch from 'node-fetch'
-
-type Issue = {
-  title: string
-  description?: string
-  estimate?: number
-}
-
-type Epic = Issue & {
-  issues: Issue[]
-}
-
-type ReleaseReport = Issue & {
-  start: string
-  end: string
-  epics: Epic[]
-  issues: Issue[]
-}
+import { generateReleaseReport, ReleaseReport } from '@celo/dev-utils/lib/release-report'
 
 const n = '5'
 const networks = ['staging', 'alfajores', 'baklava', 'mainnet']
@@ -63,109 +45,8 @@ const release_report: ReleaseReport = {
   issues: [],
 }
 
-const req = (
-  url_prefix: string,
-  headers: Record<string, string>,
-  base_query?: URLSearchParams
-) => async (method: 'GET' | 'POST' | 'PUT' | 'PATCH', url: string, params: Record<string, any>) => {
-  const options = { method, headers, body: undefined }
-  let full_url = `${url_prefix}/${url}?${(base_query ?? '').toString()}`
-
-  if (method === 'GET') {
-    full_url += new URLSearchParams(params).toString()
-  } else {
-    options.body = JSON.stringify(params)
-  }
-
-  const resp = await fetch(full_url, options)
-  if (!resp.ok) {
-    throw new Error(
-      `Request failed: ${method} ${full_url} with ${resp.status} (${resp.statusText})`
-    )
-  }
-  return JSON.parse(await resp.text())
-}
-
-const issue_labels = ['Component: Contracts', 'audit', 'CAP']
-const gh_org = 'celo-org'
-const gh_repo = 'celo-monorepo'
-
-const github_req = req(`https://api.github.com/repos/${gh_org}/${gh_repo}`, {
-  Authorization: `token ${process.env.GITHUB_PAT ?? ''}`,
-  Accept: 'application/vnd.github.v3+json',
-  'Content-Type': 'application/json',
-})
-
-const repo_id = '197642503'
-const workspace_id = '600598462807be0011921c65'
-
-const zenhub_req = req(
-  'https://api.zenhub.com/p1',
-  {
-    'X-Authentication-Token': process.env.ZENHUB_API_KEY ?? '',
-    'Content-Type': 'application/json',
-  },
-  new URLSearchParams({ repo_id, workspace_id })
-)
-
-const convertGithubToZenhubIssues = (githubIssues: Array<any>) =>
-  githubIssues.map((gh_issue) => ({
-    repo_id: parseInt(repo_id),
-    issue_number: parseInt(gh_issue.number),
-  }))
-
-const main = async (releaseReport: ReleaseReport): Promise<void> => {
-  // create release
-  const newReport = await zenhub_req('POST', `repositories/${repo_id}/reports/release`, {
-    title: releaseReport.title,
-    description: releaseReport.description,
-    start_date: new Date(releaseReport.start).toISOString(),
-    desired_end_date: new Date(releaseReport.end).toISOString(),
-    repositories: [parseInt(repo_id)],
-  })
-
-  // create epics
-  const gh_epics = await concurrentMap(1, releaseReport.epics, async (epic) => {
-    // create github issue representing epic
-    const gh_epic = await github_req('POST', 'issues', {
-      title: epic.title,
-      body: epic.description,
-      labels: issue_labels,
-    })
-
-    // create github issues representing epic contents
-    const gh_issues = await concurrentMap(1, epic.issues, (issue) =>
-      github_req('POST', 'issues', { title: issue.title, labels: issue_labels })
-    )
-    const zh_issues = convertGithubToZenhubIssues(gh_issues)
-
-    // assign estimates to github issues
-    await concurrentMap(1, zh_issues, (zh_issue, i) =>
-      zenhub_req('PUT', `repositories/${repo_id}/issues/${zh_issue.issue_number}/estimate`, {
-        estimate: epic.issues[i].estimate,
-      })
-    )
-
-    // convert github issue to epic and add contents
-    await zenhub_req('POST', `repositories/${repo_id}/issues/${gh_epic.number}/convert_to_epic`, {
-      issues: zh_issues,
-    })
-
-    // add issues to release report
-    await zenhub_req('PATCH', `reports/release/${newReport.release_id}/issues`, {
-      add_issues: zh_issues,
-      remove_issues: [],
-    })
-
-    return gh_epic
-  })
-
-  const zh_epics = convertGithubToZenhubIssues(gh_epics)
-  // add epics to release report
-  await zenhub_req('PATCH', `reports/release/${newReport.release_id}/issues`, {
-    add_issues: zh_epics,
-    remove_issues: [],
-  })
-}
-
-main(release_report).catch((error) => console.error(error))
+generateReleaseReport(
+  { labels: ['Component: Contracts', 'audit', 'CAP'], org: 'celo-org', repo: 'celo-monorepo' },
+  { repo_id: '197642503', workspace_id: '600598462807be0011921c65' },
+  release_report
+).catch((error) => console.error(error))
